Major UI Improvements
Added a last page after the quiz is over. Next button only appears when needed.
Changed visibility "hidden" to "display none" for the dynamic hiding of elements.
Changed the location of Score tab and Center aligned the question title.
Use of labels for checkboxes and radio button allows the user to click on the text to select the checkbox instead of the checkbox.